Extract watchlists service tier
