+++ apps/web/src/lib/actions/__tests__/watchlists-detail-perf.test.ts
@@ -148,6 +148,14 @@ vi.mock("@/lib/actions/taxonomy", () => ({
   resolveTechnologySlugs: vi.fn().mockResolvedValue(new Map()),
 }));
 
+vi.mock("@/lib/services/taxonomy", () => ({
+  expandOccupationIds: vi.fn().mockResolvedValue([]),
+  expandOccupationIdsBatch: vi.fn().mockResolvedValue([]),
+  resolveOccupationSlugs: vi.fn().mockResolvedValue(new Map()),
+  resolveSenioritySlugs: vi.fn().mockResolvedValue(new Map()),
+  resolveTechnologySlugs: vi.fn().mockResolvedValue(new Map()),
+}));
+
 vi.mock("drizzle-orm", () => ({
   sql: (..._args: unknown[]) => ({ _isSql: true }),
   eq: (..._args: unknown[]) => ({ _isEq: true }),
